Move Extension model to Services
