+++ hugashop/crm/Services/Extension.php
@@ -4,11 +4,11 @@
  * HugaShop - Sell anything
  *
  * @author Andri Huga
- * @version 1.3
+ * @version 1.4
  *
  */
 
-namespace HugaShop\Models;
+namespace HugaShop\Services;
 
 use HugaShop\Models\Settings;
 use HugaShop\Services\Config;
