+++ src/Controller/Admin/Extension/ExtensionController.php
@@ -4,7 +4,7 @@
  * HugaShop - Sell anything
  *
  * @author Andri Huga
- * @version 2.1
+ * @version 2.2
  *
  */
 
@@ -12,7 +12,7 @@ namespace App\Controller\Admin\Extension;
 
 use HugaShop\Services\Design;
 use HugaShop\Services\Request;
-use HugaShop\Models\Extension;
+use HugaShop\Services\Extension;
 use App\Controller\BaseAdminController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
