+++ src/Controller/Admin/Extension/ExtensionListController.php
@@ -4,14 +4,14 @@
  * HugaShop - Sell anything
  *
  * @author Andri Huga
- * @version 2.0
+ * @version 2.1
  *
  */
 
 namespace App\Controller\Admin\Extension;
 
 use HugaShop\Services\Design;
-use HugaShop\Models\Extension;
+use HugaShop\Services\Extension;
 use App\Controller\BaseAdminController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
