+++ src/Controller/Front/Exchange/ExtensionController.php
@@ -3,13 +3,13 @@
 /**
  *
  * @author Andi Huga
- * @version 1.9
+ * @version 2.0
  *
  */
 
 namespace App\Controller\Front\Exchange;
 
-use HugaShop\Models\Extension;
+use HugaShop\Services\Extension;
 use App\Controller\BaseFrontController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
